add CalendarSelector component and initial loading screen for app
