+++ src/lib/api/data.ts
@@ -17,6 +17,36 @@ export async function seedData() {
     },
   });
 
+  await calendars.create({
+    fields: {
+      name: "Pedals 3000",
+      dates: [
+        {
+          start: new Date("2025-01-20T00:00:00Z"),
+          end: new Date("2025-01-27T00:00:00Z"),
+        },
+      ],
+      festivalInstructions: null,
+      spacePageText: null,
+      resourcePageText: null
+    },
+  });
+
+  await calendars.create({
+    fields: {
+      name: "Cute venue",
+      dates: [
+        {
+          start: new Date("2025-01-20T00:00:00Z"),
+          end: new Date("2025-01-27T00:00:00Z"),
+        },
+      ],
+      festivalInstructions: null,
+      spacePageText: null,
+      resourcePageText: null
+    },
+  });
+
   await setActiveCalendar(calendarId);
 
   const spaceOneId = await spaces.create(calendarId, {
