+++ src/routes/(unauthed)/join/+page.ts
@@ -1,21 +1,17 @@
 import type { PageLoad } from './$types';
-import { access, calendars, identity } from "$lib/api";
+import { calendars } from "$lib/api";
 import { goto } from '$app/navigation';
 
 export const load: PageLoad = async () => {
-  let myPublicKey = await identity.publicKey();
-  let activeCalendarId = await calendars.getActiveCalendarId()
-  let accessStatus = await access.checkStatus(myPublicKey, activeCalendarId!);
+  let activeCalendarId = await calendars.getActiveCalendarId();
 
-  // if we already have access then go to the calendar
-  if (accessStatus == 'accepted') {
+  // TODO: check if user has access to active calendar before going to the calendar
+  // If access is pending then go to join/pending page.
+  if (activeCalendarId) {
     goto("/app/events")
   }
 
   return {
-    title: 'request access',
-    myPublicKey,
-    activeCalendarId,
-    accessStatus
+    title: 'welcome'
   };
 }
